Use an inline template with _.template to render the gallery's HTML
`<script type="text/html">` is neither rendered nor not parsed by the browser's JavaScript interpreter, so the tag may be used to store arbitrary HTML fragments such as our template's markup.
In thie case we store an Underscore.js template, which allows us to move the markup back to the document whence it came.