Break Backbone Model and View definitions into separate files
This is a small project, but in a larger application it is very helpful to work with many small files rather than one monolithic one. Finding what you want requires less scrolling, and there are many fewer opportunities for your code changes to collide with your teammate's if code is properly broken out.

I recommend using a package manager such as [Require.js](http://requirejs.org/) to handle your script dependencies; even if you use WordPress to enqueue the final file, Require can be used as a helpful build step. The main reason we are not including it here is to limit the number of moving parts in this demo -- i.e., it is hopefully easier to see how Backbone works when you aren't also trying to understand a module loader at the same time.